requirements.txt
click
rich
pyyaml
requests
python-dotenv